changes to reflect overhaul of analysis classes

+++ test/stubs/ApvTimingHistosUsingDb.h
@@ -26,12 +26,12 @@ class ApvTimingHistosUsingDb : public ApvTimingHistograms,
  private:
 
   void settings( std::map<uint32_t,uint32_t>& pll_delays,
-		 std::map<uint32_t,float>& ticker_thresh );
+		 std::map<uint32_t,uint32_t>& ticker_thresh );
   
   void update( const std::map<uint32_t,uint32_t>&,
 	       SiStripConfigDb::DeviceDescriptions& );
 
-  void update( const std::map<uint32_t,float>&,
+  void update( const std::map<uint32_t,uint32_t>&,
 	       SiStripConfigDb::FedDescriptions& );
